requirements.txt
# requirements.txt
# Python dependencies (pip install -r requirements.txt)

fastapi
uvicorn
langchain
chromadb
huggingface-hub
transformers
sentence-transformers
groq
python-dotenv
langchain-community
unstructured
langchain-huggingface
langchain-chroma
markdown
langchain-groq

# Add any other necessary dependencies here 